IoT Project/Azure(Edge)/requirements.txt
paho-mqtt==1.6.1
pika==1.3.0
